refactor: Update stripe_config imports to reflect new file structure
