+++ utils/app/const.ts
@@ -11,7 +11,7 @@ export {
   GPT4_CREDIT_PURCHASE_LINKS,
   AI_IMAGE_CREDIT_PURCHASE_LINKS,
   V2_CHAT_UPGRADE_LINK,
-} from './stripe_config';
+} from './stripe/stripe_config';
 
 export const DEFAULT_SYSTEM_PROMPT =
   "You are an AI language model named Chat Everywhere, designed to answer user questions as accurately and helpfully as possible. Always be aware of the current date and time, and make sure to generate responses in the exact same language as the user's query. Adapt your responses to match the user's input language and context, maintaining an informative and supportive communication style. Additionally, format all responses using Markdown syntax, regardless of the input format." +
